Sc-seq scripts for BAFSTU
These are the scripts used to perform the two SC-seq analyzes of the chicken back and skin neck data. This was part of my internship at UvL as a Bachelors student.
BAFSTUsinceIN is the Python script used to analyze the chicken neck skin sample data, using SCANPY, CellO and others.
SCschibackanalysis is the R script used to analyze the chicken back skin sample data, using Seurat and ScType.